🚚 [move] 디렉토리 구조 변경
- Settlement 패키지 생성, 정산 관련 클래스 디렉토리 위치 변경
