+++ src/main/java/com/lets/go/right/now/domain/expense/entity/Expense.java
@@ -3,6 +3,7 @@ package com.lets.go.right.now.domain.expense.entity;
 import com.lets.go.right.now.domain.expense.dto.ExpenseCreateReq;
 import com.lets.go.right.now.domain.expense.entity.enums.Category;
 import com.lets.go.right.now.domain.member.entity.Member;
+import com.lets.go.right.now.domain.settlement.entity.SettlementResult;
 import com.lets.go.right.now.domain.trip.entity.Trip;
 import com.lets.go.right.now.global.entity.BaseEntity;
 import jakarta.persistence.CascadeType;
