+++ src/main/java/com/lets/go/right/now/domain/expense/service/ExpenseServiceImpl.java
@@ -4,11 +4,11 @@ import com.lets.go.right.now.domain.expense.dto.ExpenseCreateReq;
 import com.lets.go.right.now.domain.expense.dto.ExpenseViewRes;
 import com.lets.go.right.now.domain.expense.entity.ExcludedMember;
 import com.lets.go.right.now.domain.expense.entity.Expense;
-import com.lets.go.right.now.domain.expense.entity.SettlementResult;
+import com.lets.go.right.now.domain.settlement.entity.SettlementResult;
 import com.lets.go.right.now.domain.expense.entity.TripImage;
 import com.lets.go.right.now.domain.expense.repository.ExcludedMemberRepository;
 import com.lets.go.right.now.domain.expense.repository.ExpenseRepository;
-import com.lets.go.right.now.domain.expense.repository.SettlementResultRepository;
+import com.lets.go.right.now.domain.settlement.repository.SettlementResultRepository;
 import com.lets.go.right.now.domain.expense.repository.TripImageRepository;
 import com.lets.go.right.now.domain.member.entity.Member;
 import com.lets.go.right.now.domain.member.repository.MemberRepository;
