+++ src/test/java/com/lets/go/right/now/domain/expense/service/ExpenseServiceImplTest.java
@@ -3,10 +3,10 @@ package com.lets.go.right.now.domain.expense.service;
 import com.lets.go.right.now.domain.expense.dto.ExpenseCreateReq;
 import com.lets.go.right.now.domain.expense.entity.ExcludedMember;
 import com.lets.go.right.now.domain.expense.entity.Expense;
-import com.lets.go.right.now.domain.expense.entity.SettlementResult;
+import com.lets.go.right.now.domain.settlement.entity.SettlementResult;
 import com.lets.go.right.now.domain.expense.repository.ExcludedMemberRepository;
 import com.lets.go.right.now.domain.expense.repository.ExpenseRepository;
-import com.lets.go.right.now.domain.expense.repository.SettlementResultRepository;
+import com.lets.go.right.now.domain.settlement.repository.SettlementResultRepository;
 import com.lets.go.right.now.domain.member.entity.Member;
 import com.lets.go.right.now.domain.member.repository.MemberRepository;
 import com.lets.go.right.now.domain.trip.entity.Trip;
@@ -17,7 +17,6 @@ import java.io.IOException;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
